Created folder src/model moved all ts-files there except index.ts and update dependencies everywhere.
