+++ frontend/src/components/ChatInteraction.vue
@@ -15,8 +15,8 @@
 <script lang="ts">
 import Vue from "vue";
 import QuestionPack from "./QuestionPack.vue";
-import {Question} from "./../Question";
-import {Answer} from "./../Answer"
+import {Question} from "./../model/Question";
+import {Answer} from "./../model/Answer"
 import TheHeader from './TheHeader.vue';
 export default Vue.extend({
     data() {
