+++ frontend/src/components/TheResultSite.vue
@@ -24,8 +24,6 @@
 
 <script lang="ts">
 import Vue from "vue";
-import { Question } from "./../Question";
-import { Answer } from "./../Answer";
 import TheHeader from "./TheHeader.vue";
 
 export default Vue.extend({
