+++ frontend/src/model/Edge.ts
