+++ frontend/src/model/Graph.ts
