+++ frontend/src/model/GraphFactory.ts
