+++ frontend/src/model/JSONEdge.ts
