+++ frontend/src/model/JSONGraph.ts
