+++ frontend/src/model/JSONNode.ts
