+++ frontend/src/model/Node.ts
