+++ frontend/tests/components/ChatBox.spec.ts
@@ -2,8 +2,8 @@ import { mount, shallowMount, Wrapper } from '@vue/test-utils'
 import Vue from "vue"
 import QuestionPack from "./../../src/components/QuestionPack.vue"
 import QuestionBox from "./../../src/components/QuestionBox.vue"
-import { Question } from "./../../src/Question"
-import { Answer } from '../../src/Answer'
+import { Question } from "./../../src/model/Question"
+import { Answer } from '../../src/model/Answer'
 
 describe("QuestionPack", () => {
   let wrapper: Wrapper<Vue>;
